* Made fixes to get AngelScript working in the Lambda Lilac Engine. * Changed entities from structures to numbers. * Added a different visualisation mode to the default shader. * Added a very simple CPU profiler.

+++ src/engine/systems/entity_system.h
@@ -21,9 +21,9 @@ namespace lambda
       virtual void deinitialize() override {};
 
     private:
-      static const size_t kFreeIdIncrement = 64u;
-      size_t free_id_count_ = 1;
-      Queue<size_t> free_ids_;
+      static const uint32_t kFreeIdIncrement = 64u;
+			uint32_t free_id_count_ = 1;
+      Queue<uint32_t> free_ids_;
     };
   }
 }
